nginx/requirements.txt
pyelasticsearch
jython
GeoIP
python-geoip
python-geoip-geolite2
